+++ js/main.js
@@ -1,10 +1,2 @@
-/* eslint-disable no-console */
 import './utils.js';
-// eslint-disable-next-line no-unused-vars
-import {photos} from './data.js';
-// eslint-disable-next-line no-unused-vars
-import {pictureElement}  from './miniature.js';
-
-// eslint-disable-next-line no-console
-//console.log(photos);
-//console.log (pictureElement);
+import './mini-pictures.js';
